#28: refactor: improve naming
