+++ src/test/resources-its/io/github/georgberky/maven/plugins/depsupdate/UpdateMojoJGitIT/jgitProviderIsSet_scmConfigIsSet/pom.xml
