+++ src/test/resources-its/io/github/georgberky/maven/plugins/depsupdate/UpdateMojoJGitScmConfigIsMissingIT/jgitProviderIsSet_scmConfigIsMissing/pom.xml
